# Exact score based algorithm optimizations